+++ core/src/test/java/cucumber/runtime/xstream/ConvertersTest.java
@@ -2,6 +2,7 @@ package cucumber.runtime.xstream;
 
 import cucumber.deps.com.thoughtworks.xstream.converters.ConverterLookup;
 import cucumber.deps.com.thoughtworks.xstream.converters.SingleValueConverter;
+import org.junit.Before;
 import org.junit.Test;
 
 import java.math.BigDecimal;
@@ -12,28 +13,30 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 public class ConvertersTest {
-    @Test
-    public void shouldTransformToTheRightType() {
+    private ConverterLookup en;
+    private ConverterLookup no;
+
+    @Before
+    public void setUp() throws Exception {
         ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
         LocalizedXStreams transformers = new LocalizedXStreams(classLoader);
+        en = transformers.get(Locale.US).getConverterLookup();
+        no = transformers.get(new Locale("no")).getConverterLookup();
+    }
 
-        ConverterLookup en = transformers.get(Locale.US).getConverterLookup();
+    @Test
+    public void shouldTransformToTheRightType() {
         assertTrue((Boolean) ((SingleValueConverter) en.lookupConverterForType(Boolean.class)).fromString("true"));
         assertTrue((Boolean) ((SingleValueConverter) en.lookupConverterForType(Boolean.TYPE)).fromString("true"));
         assertEquals(3000.15f, (Float) ((SingleValueConverter) en.lookupConverterForType(Float.class)).fromString("3000.15"), 0.000001);
         assertEquals(3000.15f, (Float) ((SingleValueConverter) en.lookupConverterForType(Float.TYPE)).fromString("3000.15"), 0.000001);
         assertEquals(new BigDecimal("3000.15"), ((SingleValueConverter) en.lookupConverterForType(BigDecimal.class)).fromString("3000.15"));
 
-        ConverterLookup no = transformers.get(new Locale("no")).getConverterLookup();
         assertEquals(3000.15f, (Float) ((SingleValueConverter) no.lookupConverterForType(Float.TYPE)).fromString("3000,15"), 0.000001);
     }
 
     @Test
     public void shouldTransformPatternWithFlags() {
-        ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
-        LocalizedXStreams transformers = new LocalizedXStreams(classLoader);
-
-        ConverterLookup en = transformers.get(Locale.US).getConverterLookup();
         Pattern expected = Pattern.compile("hello", Pattern.CASE_INSENSITIVE | Pattern.MULTILINE);
         Pattern actual = (Pattern) ((SingleValueConverter) en.lookupConverterForType(Pattern.class)).fromString("/hello/im");
         assertEquals(expected.pattern(), actual.pattern());
@@ -42,10 +45,6 @@ public class ConvertersTest {
 
     @Test
     public void shouldTransformPatternWithoutFlags() {
-        ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
-        LocalizedXStreams transformers = new LocalizedXStreams(classLoader);
-
-        ConverterLookup en = transformers.get(Locale.US).getConverterLookup();
         Pattern expected = Pattern.compile("hello");
         Pattern actual = (Pattern) ((SingleValueConverter) en.lookupConverterForType(Pattern.class)).fromString("hello");
         assertEquals(expected.pattern(), actual.pattern());
@@ -54,10 +53,6 @@ public class ConvertersTest {
 
     @Test
     public void shouldIncludeSlashesInPatternWhenThereAreNoFlags() {
-        ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
-        LocalizedXStreams transformers = new LocalizedXStreams(classLoader);
-
-        ConverterLookup en = transformers.get(Locale.US).getConverterLookup();
         Pattern expected = Pattern.compile("/hello/");
         Pattern actual = (Pattern) ((SingleValueConverter) en.lookupConverterForType(Pattern.class)).fromString("/hello/");
         assertEquals(expected.pattern(), actual.pattern());
@@ -66,18 +61,12 @@ public class ConvertersTest {
 
     @Test
     public void shouldTransformToTypeWithStringCtor() {
-        ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
-        LocalizedXStreams transformers = new LocalizedXStreams(classLoader);
-        ConverterLookup en = transformers.get(Locale.US).getConverterLookup();
         SingleValueConverter c = (SingleValueConverter) en.lookupConverterForType(MyClass.class);
         assertEquals("X", ((MyClass) c.fromString("X")).s);
     }
 
     @Test
     public void shouldTransformToTypeWithObjectCtor() {
-        ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
-        LocalizedXStreams transformers = new LocalizedXStreams(classLoader);
-        ConverterLookup en = transformers.get(Locale.US).getConverterLookup();
         SingleValueConverter c = (SingleValueConverter) en.lookupConverterForType(MyOtherClass.class);
         assertEquals("X", ((MyOtherClass) c.fromString("X")).o);
     }